Split out stack files
